requirements.txt
mkdocs-material==6.1.4
mkdocs-git-revision-date-localized-plugin==0.7.3
mkdocs-git-authors-plugin==0.3.2
mkdocs-rss-plugin==0.9.0
gitpython==3.1.11
dominate==2.6.0
requests==2.24.0
